Programa - Primer Commit

+++ src/main/java/com/example/demo/repository/AttendanceRepository.java
@@ -1,7 +1,8 @@
 package com.example.demo.repository;
 
-import com.example.demo.model.Attendance;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.example.demo.model.Attendance;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
+
 }
